requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg[binary]==3.2.6
pydantic==2.11.3
pydantic-settings==2.8.1
requests==2.32.3
langgraph==0.2.60
langchain-core==0.3.28
openai==1.76.2
google-generativeai==0.8.5
python-dotenv==1.1.0
pytest==8.3.5
httpx==0.28.1
streamlit==1.45.1
